refactor: update temperature display for int representation

+++ lib/Display/Display.h
@@ -25,6 +25,6 @@ void clearPlot();
 void drawIcons();
 void writeTime(struct tm time, uint8_t x, uint8_t y, uint8_t w, uint8_t h, uint8_t size);
 void writeData(struct tm time, uint8_t x, uint8_t y, uint8_t w, uint8_t h, uint8_t size);
-void writeTemperature(float value, uint8_t x, uint8_t y, uint8_t w, uint8_t h, uint8_t size);
+void writeTemperature(int16_t value, uint8_t x, uint8_t y, uint8_t w, uint8_t h, uint8_t size);
 uint8_t drawMarker(int16_t move);
 void drawHistoryTemplate();
